Linked List-Linked List Construction
create a class node, then make constructor for it. then make 2 vars head and tail. then make a method to add a node. then another method to display the output. and then main method where u print them.
In the method to add a node, this is imp
if(head == null) {
            head = newN;
            tail = newN;
        }
        else {
            tail.next = newN;
            tail = newN;
        }
.......